возврат фул адс дто

+++ src/main/java/com/example/finaltask/model/dto/FullAdsDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class FullAdsDTO {
-    private Long pk;
+    private Integer pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
